1318 - move index folder to client
