+++ src/components/client/index/IndexPage.styled.tsx
