+++ src/components/client/index/sections/FaqSection/FaqSection.styled.tsx
