+++ src/components/client/index/sections/HowWeWorkSection/HowWeWorkSection.styled.tsx
