+++ src/components/client/index/sections/JoinPodkrepiBgSection/JoinPodkrepiBgSection.styled.tsx
